requirements.txt
django-ckeditor==6.4.0